+++ backend/Application/Api/GraphQL/GraphQlConfiguration.cs
@@ -2,6 +2,8 @@
 using Application.Api.GraphQL.Bakers;
 using Application.Api.GraphQL.Blocks;
 using Application.Api.GraphQL.Contracts;
+using Application.Api.GraphQL.ChainParametersGraphql;
+using Application.Api.GraphQL.Import;
 using Application.Api.GraphQL.Metrics;
 using Application.Api.GraphQL.Network;
 using Application.Api.GraphQL.Pagination;
@@ -54,6 +56,8 @@ public static class GraphQlConfiguration
             .AddType<PoolRewardMetricsQuery>()
             .AddType<PassiveDelegationQuery>()
             .AddType<PaydayQuery>()
+            .AddType<ChainParametersQuery>()
+            .AddType<ImportStateQuery>()
             .AddType<NetworkQuery>();
         
         builder.AddSubscriptionType<Subscription>();