+++ backend/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Handling of block `1385825` on testnet. Which changes the NextPaydayTime by just 287 seconds.
 - Added query to fetch backend version.
 - Sorting on Nodes Page
+- Fixed Issue 5 by considering 0 balance for Newly created accounts.
 - Added boolean filters on Bakers Query to enable filtering on Baker Removed State.
 - Fixed Memo CBOR Decoding.
 - Adds Listing of Contracts Via View